Change standardization neritic vs oceanic
Standardization was done per element across all areas, which does not make sense given we want to compare levels of nutrient release between habitat within each area. So I change to a standardization per element and per area! Standardization selecting only values in between the lowest and highest quantiles, and using these quantiles as min and max in the standardization formula. This avoided to get super shreded standardized values due to the extreme max and min accounting for only a few extreme estimates.